fix: adds wallet standard support to orbit-solana
Adds support for Wallet Standard, allowing for a standardized way to access Solana wallets.
This includes adding dependencies, exporting new utility functions for wallet interactions,
and updating the Solana adapter to use the new wallet standard functions for fetching wallet information such as name and avatar.
Also fixes an issue where the cluster parameter was not being correctly passed to the explorer link.